[qa] Split free_radius views and test_api #196
Closes #196
