+++ openwisp_radius/tests/test_api/test_registration.py
@@ -2,7 +2,6 @@ import json
 import os
 import sys
 import uuid
-from datetime import timedelta
 from unittest import mock
 
 import swapper
@@ -13,43 +12,38 @@ from django.contrib.auth.models import Permission
 from django.contrib.auth.tokens import default_token_generator
 from django.core import mail
 from django.core.cache import cache
-from django.core.exceptions import ValidationError
 from django.core.mail import EmailMultiAlternatives
 from django.test import override_settings
 from django.urls import reverse
 from django.utils.crypto import get_random_string
 from django.utils.encoding import force_bytes, force_text
 from django.utils.http import urlsafe_base64_encode
-from django.utils.timezone import localtime, now
+from django.utils.timezone import now
 from freezegun import freeze_time
 from rest_framework import status
 from rest_framework.authtoken.models import Token
 from rest_framework.test import APIClient
 
-from openwisp_radius.api import views as api_views
 from openwisp_utils.tests import capture_any_output, capture_stderr
 
-from .. import settings as app_settings
-from ..utils import load_model
-from . import _TEST_DATE, FileMixin
-from .mixins import ApiTokenMixin, BaseTestCase
+from ... import settings as app_settings
+from ...utils import load_model
+from ..mixins import ApiTokenMixin, BaseTestCase
 
 User = get_user_model()
-PhoneToken = load_model('PhoneToken')
 RadiusToken = load_model('RadiusToken')
+RadiusBatch = load_model('RadiusBatch')
 RadiusAccounting = load_model('RadiusAccounting')
 RadiusPostAuth = load_model('RadiusPostAuth')
-RadiusBatch = load_model('RadiusBatch')
-RadiusUserGroup = load_model('RadiusUserGroup')
 OrganizationRadiusSettings = load_model('OrganizationRadiusSettings')
 Organization = swapper.load_model('openwisp_users', 'Organization')
-OrganizationUser = swapper.load_model('openwisp_users', 'OrganizationUser')
 
 START_DATE = '2019-04-20T22:14:09+01:00'
 
 
-class TestApi(ApiTokenMixin, FileMixin, BaseTestCase):
+class TestApi(ApiTokenMixin, BaseTestCase):
     _test_email = 'test@openwisp.org'
+    _acct_url = reverse('radius:accounting')
 
     def setUp(self):
         cache.clear()
@@ -1292,7 +1286,7 @@ class TestApi(ApiTokenMixin, FileMixin, BaseTestCase):
         self.assertEqual(RadiusBatch.objects.count(), 0)
         self.assertEqual(User.objects.count(), 0)
         path_csv = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)), 'static', 'test_batch.csv'
+            os.path.dirname(os.path.dirname(__file__)), 'static', 'test_batch.csv'
         )
         with open(path_csv, 'rt') as file:
             data = self._radius_batch_csv_data(csvfile=file)
@@ -1671,1178 +1665,3 @@ class TestApi(ApiTokenMixin, FileMixin, BaseTestCase):
                 },
             )
             self.assertEqual(r.status_code, 403)
-
-
-class TestApiReject(ApiTokenMixin, BaseTestCase):
-    @classmethod
-    def setUpClass(cls):
-        app_settings.API_AUTHORIZE_REJECT = True
-
-    @classmethod
-    def tearDownClass(cls):
-        app_settings.API_AUTHORIZE_REJECT = False
-
-    def test_disabled_user_login(self):
-        User.objects.create_user(username='barbar', password='molly', is_active=False)
-        response = self.client.post(
-            reverse('radius:authorize'),
-            {'username': 'barbar', 'password': 'molly'},
-            HTTP_AUTHORIZATION=self.auth_header,
-        )
-        self.assertEqual(response.status_code, 401)
-        self.assertEqual(response.data, {'control:Auth-Type': 'Reject'})
-
-    def test_authorize_401(self):
-        response = self.client.post(
-            reverse('radius:authorize'),
-            {'username': 'baldo', 'password': 'ugo'},
-            HTTP_AUTHORIZATION=self.auth_header,
-        )
-        self.assertEqual(response.status_code, 401)
-        self.assertEqual(response.data, {'control:Auth-Type': 'Reject'})
-
-
-class TestAutoGroupname(ApiTokenMixin, BaseTestCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        app_settings.API_ACCOUNTING_AUTO_GROUP = True
-
-    def test_automatic_groupname_account_enabled(self):
-        user = User.objects.create_superuser(
-            username='username1', email='admin@admin.com', password='qwertyuiop'
-        )
-        usergroup1 = self._create_radius_usergroup(
-            groupname='group1', priority=2, username='testgroup1'
-        )
-        usergroup2 = self._create_radius_usergroup(
-            groupname='group2', priority=1, username='testgroup2'
-        )
-        user.radiususergroup_set.set([usergroup1, usergroup2])
-        self.client.post(
-            f'{reverse("radius:accounting")}{self.token_querystring}',
-            {
-                'status_type': 'Start',
-                'session_time': '',
-                'input_octets': '',
-                'output_octets': '',
-                'nas_ip_address': '127.0.0.1',
-                'session_id': '48484',
-                'unique_id': '1515151',
-                'username': 'username1',
-            },
-        )
-        accounting_created = RadiusAccounting.objects.get(username='username1')
-        self.assertEqual(accounting_created.groupname, 'group2')
-        user.delete()
-
-
-class TestAutoGroupnameDisabled(ApiTokenMixin, BaseTestCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        app_settings.API_ACCOUNTING_AUTO_GROUP = False
-
-    def test_account_creation_api_automatic_groupname_disabled(self):
-        user = User.objects.create_superuser(
-            username='username1', email='admin@admin.com', password='qwertyuiop'
-        )
-        usergroup1 = self._create_radius_usergroup(
-            groupname='group1', priority=2, username='testgroup1'
-        )
-        usergroup2 = self._create_radius_usergroup(
-            groupname='group2', priority=1, username='testgroup2'
-        )
-        user.radiususergroup_set.set([usergroup1, usergroup2])
-        url = f'{reverse("radius:accounting")}{self.token_querystring}'
-        self.client.post(
-            url,
-            {
-                'status_type': 'Start',
-                'session_time': '',
-                'input_octets': '',
-                'output_octets': '',
-                'nas_ip_address': '127.0.0.1',
-                'session_id': '48484',
-                'unique_id': '1515151',
-                'username': 'username1',
-            },
-        )
-        accounting_created = RadiusAccounting.objects.get(username='username1')
-        self.assertIsNone(accounting_created.groupname)
-        user.delete()
-
-
-class TestApiUserToken(ApiTokenMixin, BaseTestCase):
-    def _get_url(self):
-        return reverse('radius:user_auth_token', args=[self.default_org.slug])
-
-    def _user_auth_token_helper(self, username):
-        url = self._get_url()
-        response = self.client.post(url, {'username': 'tester', 'password': 'tester'})
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data['key'], Token.objects.first().key)
-        self.assertEqual(
-            response.data['radius_user_token'], RadiusToken.objects.first().key,
-        )
-        self.assertTrue(response.data['is_active'])
-
-    def test_user_auth_token_200(self):
-        org_user = self._get_org_user()
-
-        with self.subTest('login with username'):
-            self._user_auth_token_helper(org_user.user.username)
-
-        with self.subTest('login with email'):
-            self._user_auth_token_helper(org_user.user.email)
-
-        with self.subTest('login with phone_number'):
-            org_user.user.phone_number = '+23767778243'
-            org_user.save()
-            self._user_auth_token_helper(org_user.user.phone_number)
-
-    def test_user_auth_token_with_second_organization(self):
-        api_views.renew_required = False
-        self._get_org_user()
-        self.org2 = self._create_org(**{'name': 'test', 'slug': 'test'})
-        self._create_org_user(**{'organization': self.org2})
-        for org_slug in [self.default_org.slug, self.org2.slug]:
-            with self.subTest(org_slug):
-                # Get token for organizations
-                response = self.client.post(
-                    reverse('radius:user_auth_token', args=[org_slug]),
-                    {'username': 'tester', 'password': 'tester'},
-                )
-                self.assertEqual(response.status_code, 200)
-                # Check authorization accepts for both organizations
-                response = self.client.post(
-                    reverse('radius:authorize'),
-                    {'username': 'tester', 'password': 'tester'},
-                )
-                self.assertEqual(response.data, {'control:Auth-Type': 'Accept'})
-        api_views.renew_required = True
-
-    def test_user_auth_token_400_credentials(self):
-        url = self._get_url()
-        r = self.client.post(url, {'username': 'tester', 'password': 'tester'})
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('Unable to log in', r.json()['non_field_errors'][0])
-
-    @capture_any_output()
-    def test_user_auth_token_400_organization(self):
-        url = self._get_url()
-        self._get_org_user()
-        OrganizationUser.objects.all().delete()
-        r = self.client.post(url, {'username': 'tester', 'password': 'tester'})
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('is not member', r.json()['non_field_errors'][0])
-
-    def test_user_auth_token_404(self):
-        url = reverse(
-            'radius:user_auth_token', args=['00000000-0000-0000-0000-000000000000']
-        )
-        r = self.client.post(url, {'username': 'tester', 'password': 'tester'})
-        self.assertEqual(r.status_code, 404)
-
-    @freeze_time(_TEST_DATE)
-    def test_user_auth_updates_last_login(self):
-        admin = self._get_admin()
-        login_payload = {'username': 'admin', 'password': 'tester'}
-        login_url = reverse('radius:user_auth_token', args=[self.default_org.slug])
-        self.assertEqual(admin.last_login, None)
-        response = self.client.post(login_url, data=login_payload)
-        self.assertEqual(response.status_code, 200)
-        admin.refresh_from_db()
-        self.assertIsNotNone(admin.last_login)
-        self.assertEqual(localtime(admin.last_login).isoformat(), _TEST_DATE)
-
-    def test_user_auth_token_inactive_user(self):
-        url = self._get_url()
-        organization_user = self._get_org_user()
-        organization_user.user.is_active = False
-        organization_user.user.save()
-        response = self.client.post(url, {'username': 'tester', 'password': 'tester'})
-        self.assertEqual(response.status_code, 401)
-        self.assertFalse(response.data['is_active'])
-
-
-class TestClientIpApi(ApiTokenMixin, BaseTestCase):
-    def setUp(self):
-        self._get_org_user()
-        self.params = {'username': 'tester', 'password': 'tester'}
-        self.fail_msg = (
-            'Request rejected: Client IP address (127.0.0.1) is not in '
-            'the list of IP addresses allowed to consume the freeradius API.'
-        )
-        self.freeradius_hosts_path = 'openwisp_radius.settings.FREERADIUS_ALLOWED_HOSTS'
-        self.client.post(
-            reverse('radius:user_auth_token', args=[self._get_org().slug]), self.params
-        )
-
-    def test_cache(self):
-        def authorize_and_asset(numOfQueries, ip_list):
-            with self.assertNumQueries(numOfQueries):
-                response = self.client.post(reverse('radius:authorize'), self.params)
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.data, {'control:Auth-Type': 'Accept'})
-            self.assertEqual(cache.get(f'ip-{org.pk}'), ip_list)
-
-        cache.clear()
-        org = self._get_org()
-        self.assertEqual(cache.get(f'ip-{org.pk}'), None)
-        with self.subTest('Without Cache'):
-            authorize_and_asset(4, [])
-        with self.subTest('With Cache'):
-            authorize_and_asset(2, [])
-        with self.subTest('Organization Settings Updated'):
-            radsetting = OrganizationRadiusSettings.objects.get(organization=org)
-            radsetting.freeradius_allowed_hosts = '127.0.0.1,192.0.2.0'
-            radsetting.save()
-            authorize_and_asset(2, ['127.0.0.1', '192.0.2.0'])
-        with self.subTest('Cache Deleted'):
-            cache.clear()
-            authorize_and_asset(4, ['127.0.0.1', '192.0.2.0'])
-
-    def test_ip_from_setting_valid(self):
-        response = self.client.post(reverse('radius:authorize'), self.params)
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data, {'control:Auth-Type': 'Accept'})
-
-    @capture_any_output()
-    def test_ip_from_setting_invalid(self):
-        test_fail_msg = (
-            'Request rejected: (localhost) in organization settings or '
-            'settings.py is not a valid IP address. Please contact administrator.'
-        )
-        with mock.patch(self.freeradius_hosts_path, ['localhost']):
-            response = self.client.post(reverse('radius:authorize'), self.params)
-        self.assertEqual(response.status_code, 403)
-        self.assertEqual(response.data['detail'], test_fail_msg)
-
-    def test_ip_from_radsetting_valid(self):
-        with mock.patch(self.freeradius_hosts_path, []):
-            radsetting = OrganizationRadiusSettings.objects.get(
-                organization=self._get_org()
-            )
-        radsetting.freeradius_allowed_hosts = '127.0.0.1'
-        radsetting.save()
-        response = self.client.post(reverse('radius:authorize'), self.params)
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data, {'control:Auth-Type': 'Accept'})
-
-    @capture_any_output()
-    def test_ip_from_radsetting_invalid(self):
-        test_fail_msg = (
-            'Request rejected: (127.0.0.500) in organization settings or '
-            'settings.py is not a valid IP address. Please contact administrator.'
-        )
-        radsetting = OrganizationRadiusSettings.objects.get(
-            organization=self._get_org()
-        )
-        radsetting.freeradius_allowed_hosts = '127.0.0.500'
-        radsetting.save()
-        with mock.patch(self.freeradius_hosts_path, []):
-            response = self.client.post(reverse('radius:authorize'), self.params)
-        self.assertEqual(response.status_code, 403)
-        self.assertEqual(response.data['detail'], test_fail_msg)
-
-    @capture_any_output()
-    def test_ip_from_radsetting_not_exist(self):
-        org2 = self._create_org(**{'name': 'test', 'slug': 'test'})
-        self._create_org_user(**{'organization': org2})
-        self.client.post(reverse('radius:user_auth_token', args=[org2.slug]))
-        with self.subTest('FREERADIUS_ALLOWED_HOSTS is 127.0.0.1'):
-            response = self.client.post(reverse('radius:authorize'), self.params)
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.data, {'control:Auth-Type': 'Accept'})
-        with self.subTest('Empty Settings'), mock.patch(self.freeradius_hosts_path, []):
-            response = self.client.post(reverse('radius:authorize'), self.params)
-            self.assertEqual(response.status_code, 403)
-            self.assertEqual(response.data['detail'], self.fail_msg)
-
-
-class TestApiValidateToken(ApiTokenMixin, BaseTestCase):
-    def _get_url(self):
-        return reverse('radius:validate_auth_token', args=[self.default_org.slug])
-
-    def _test_validate_auth_token_helper(self, user):
-        url = self._get_url()
-        token = Token.objects.create(user=user)
-        # empty payload
-        response = self.client.post(url)
-        self.assertEqual(response.status_code, 401)
-        self.assertEqual(response.data['response_code'], 'BLANK_OR_INVALID_TOKEN')
-        # invalid token
-        payload = dict(token='some-random-string')
-        response = self.client.post(url, payload)
-        self.assertEqual(response.status_code, 401)
-        self.assertEqual(response.data['response_code'], 'BLANK_OR_INVALID_TOKEN')
-        # valid token
-        payload = dict(token=token.key)
-        response = self.client.post(url, payload)
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(
-            response.data['response_code'], 'AUTH_TOKEN_VALIDATION_SUCCESSFUL'
-        )
-        self.assertEqual(response.data['auth_token'], token.key)
-        self.assertEqual(
-            response.data['radius_user_token'], RadiusToken.objects.first().key,
-        )
-        user = token.user
-        self.assertEqual(user, RadiusToken.objects.first().user)
-        self.assertEqual(
-            response.data['username'], user.username,
-        )
-        self.assertEqual(
-            response.data['is_active'], user.is_active,
-        )
-        if user.is_active:
-            phone_number = user.phone_number
-        else:
-            phone_number = PhoneToken.objects.filter(user=user).first().phone_number
-
-        self.assertEqual(
-            response.data['phone_number'], str(phone_number),
-        )
-        self.assertEqual(
-            response.data['email'], user.email,
-        )
-
-    def test_validate_auth_token_with_active_user(self):
-        user = self._get_user_with_org()
-        self._test_validate_auth_token_helper(user)
-
-    @capture_any_output()
-    def test_validate_auth_token_with_inactive_user(self):
-        user = self._get_user_with_org()
-        user.is_active = False
-        user.save()
-        user.refresh_from_db()
-        phone_token = PhoneToken(
-            user=user, ip='127.0.0.1', phone_number='+237675578296'
-        )
-        phone_token.full_clean()
-        phone_token.save()
-        self._test_validate_auth_token_helper(user)
-
-    @freeze_time(_TEST_DATE)
-    def test_user_auth_updates_last_login(self):
-        admin = self._get_admin()
-        token = Token.objects.create(user=admin)
-        payload = dict(token=token.key)
-        self.assertEqual(admin.last_login, None)
-        response = self.client.post(self._get_url(), payload)
-        self.assertEqual(response.status_code, 200)
-        admin.refresh_from_db()
-        self.assertIsNotNone(admin.last_login)
-        self.assertEqual(localtime(admin.last_login).isoformat(), _TEST_DATE)
-
-
-class TestOgranizationRadiusSettings(ApiTokenMixin, BaseTestCase):
-    def setUp(self):
-        cache.clear()
-        self.org = self._create_org(**{'name': 'test', 'slug': 'test'})
-
-    def test_string_representation(self):
-        rad = OrganizationRadiusSettings.objects.create(organization=self.org)
-        self.assertEqual(str(rad), rad.organization.name)
-
-    def test_default_token(self):
-        rad = OrganizationRadiusSettings.objects.create(organization=self.org)
-        self.assertEqual(32, len(rad.token))
-
-    def test_bad_token(self):
-        try:
-            rad = OrganizationRadiusSettings(
-                token='bad.t.o.k.e.n', organization=self.org
-            )
-            rad.full_clean()
-        except ValidationError as e:
-            self.assertEqual(
-                e.message_dict['token'][0],
-                'This value must not contain spaces, dots or slashes.',
-            )
-
-    def test_cache(self):
-        rad = OrganizationRadiusSettings.objects.create(
-            token='12345', organization=self.org
-        )
-        self._get_org_user()
-        token_querystring = f'?token={rad.token}&uuid={str(self.org.pk)}'
-        post_url = f'{reverse("radius:authorize")}{token_querystring}'
-        # Clear cache before sending request
-        cache.clear()
-        self.client.post(post_url, {'username': 'tester', 'password': 'tester'})
-        with self.subTest('Cache & token match'):
-            self.assertEqual(rad.token, cache.get(rad.organization.pk))
-        with self.subTest('Force changed corrupt cache returns 403'):
-            cache.set(rad.organization.pk, 'wrong-value')
-            response = self.client.post(
-                post_url, {'username': 'tester', 'password': 'tester'}
-            )
-            self.assertEqual(response.status_code, 403)
-        with self.subTest('Updated cache with radius setting token'):
-            rad.token = '1234567'
-            rad.save()
-            self.assertEqual(rad.token, cache.get(rad.organization.pk))
-        with self.subTest('Cache Miss: radius settings deleted'):
-            rad.delete()
-            self.assertEqual(None, cache.get(rad.organization.pk))
-
-    def test_no_org_radius_setting(self):
-        self._get_org_user()
-        token_querystring = f'?token=12345&uuid={str(self.org.pk)}'
-        post_url = f'{reverse("radius:authorize")}{token_querystring}'
-        r = self.client.post(post_url, {'username': 'tester', 'password': 'tester'})
-        self.assertEqual(r.status_code, 403)
-        self.assertEqual(r.data, {'detail': 'Token authentication failed'})
-
-    def test_uuid_in_cache(self):
-        rad = OrganizationRadiusSettings.objects.create(
-            token='12345', organization=self.org
-        )
-        cache.set('uuid', str(self.org.pk), 30)
-        self._get_org_user()
-        token_querystring = f'?token={rad.token}&uuid={str(self.org.pk)}'
-        post_url = f'{reverse("radius:authorize")}{token_querystring}'
-        r = self.client.post(post_url, {'username': 'tester', 'password': 'tester'})
-        self.assertEqual(r.status_code, 200)
-
-    def test_default_organisation_radius_settings(self):
-        org = self._get_org()
-        self.assertTrue(hasattr(org, 'radius_settings'))
-        self.assertIsInstance(org.radius_settings, OrganizationRadiusSettings)
-
-    def test_sms_phone_required(self):
-        radius_settings = OrganizationRadiusSettings(
-            organization=self.org, sms_verification=True, sms_sender=''
-        )
-        try:
-            radius_settings.full_clean()
-        except ValidationError as e:
-            self.assertIn('sms_sender', e.message_dict)
-        else:
-            self.fail('ValidationError not raised')
-
-
-class TestApiPhoneToken(ApiTokenMixin, BaseTestCase):
-
-    _test_email = 'test@openwisp.org'
-
-    def setUp(self):
-        super().setUp()
-        radius_settings = self.default_org.radius_settings
-        radius_settings.sms_verification = True
-        radius_settings.sms_sender = '+595972157632'
-        radius_settings.sms_meta_data = {
-            'clientId': 3,
-            'clientSecret': 'p8pVUoqz7T4ArPEu1rcgzlsYJmnGCkkWHs2WAqqm',
-        }
-        radius_settings.full_clean()
-        radius_settings.save()
-
-    def test_register_phone_required(self):
-        self.assertEqual(User.objects.count(), 0)
-        url = reverse('radius:rest_register', args=[self.default_org.slug])
-        r = self.client.post(
-            url,
-            {
-                'username': self._test_email,
-                'email': self._test_email,
-                'password1': 'password',
-                'password2': 'password',
-                'phone_number': '',
-            },
-        )
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('phone_number', r.data)
-
-    def test_register_201(self):
-        # ensure session authentication does not interfere with the API
-        # otherwise users being logged in the admin and testing the API
-        # will get failures. Rather than telling them to log out from the admin
-        # we'll handle this case and avoid the issue altogether
-        self._superuser_login()
-        self.assertEqual(User.objects.count(), 1)
-        url = reverse('radius:rest_register', args=[self.default_org.slug])
-        phone_number = '+393664255801'
-        r = self.client.post(
-            url,
-            {
-                'username': self._test_email,
-                'email': self._test_email,
-                'password1': 'password',
-                'password2': 'password',
-                'phone_number': phone_number,
-            },
-        )
-        self.assertEqual(r.status_code, 201)
-        self.assertIn('key', r.data)
-        self.assertEqual(User.objects.count(), 2)
-        user = User.objects.get(email=self._test_email)
-        self.assertTrue(user.is_member(self.default_org))
-        self.assertEqual(user.phone_number, phone_number)
-        self.assertFalse(user.is_active)
-
-    def test_register_400_duplicate_phone_number(self):
-        self.test_register_201()
-        url = reverse('radius:rest_register', args=[self.default_org.slug])
-        r = self.client.post(
-            url,
-            {
-                'username': 'test2@test.org',
-                'email': 'test2@test.org',
-                'password1': 'password',
-                'password2': 'password',
-                'phone_number': '+393664255801',
-            },
-        )
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('phone_number', r.data)
-        self.assertEqual(User.objects.count(), 2)
-
-    def test_create_phone_token_401(self):
-        url = reverse('radius:phone_token_create', args=[self.default_org.slug])
-        r = self.client.post(url)
-        self.assertEqual(r.status_code, 401)
-
-    @mock.patch('openwisp_radius.utils.SmsMessage.send')
-    def test_create_phone_token_201(self, send_messages_mock):
-        self.test_register_201()
-        token = Token.objects.last()
-        url = reverse('radius:phone_token_create', args=[self.default_org.slug])
-        r = self.client.post(url, HTTP_AUTHORIZATION=f'Bearer {token.key}')
-        self.assertEqual(r.status_code, 201)
-        phonetoken = PhoneToken.objects.first()
-        self.assertIsNotNone(phonetoken)
-        send_messages_mock.assert_called_once()
-
-    @capture_any_output()
-    def test_create_phone_token_400_not_member(self):
-        self.test_register_201()
-        OrganizationUser.objects.all().delete()
-        token = Token.objects.last()
-        url = reverse('radius:phone_token_create', args=[self.default_org.slug])
-        r = self.client.post(url, HTTP_AUTHORIZATION=f'Bearer {token.key}')
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('non_field_errors', r.data)
-        self.assertIn('is not member', str(r.data['non_field_errors']))
-
-    @capture_any_output()
-    def test_create_phone_token_400_validation_error(self):
-        user = self._get_user_with_org()
-        user.is_active = False
-        user.save()
-        token = Token.objects.create(user=user)
-        url = reverse('radius:phone_token_create', args=[self.default_org.slug])
-        r = self.client.post(url, HTTP_AUTHORIZATION=f'Bearer {token.key}')
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('phone_number', r.data)
-        self.assertIn('This field cannot be null.', str(r.data['phone_number']))
-
-    @capture_any_output()
-    def test_create_phone_token_201_user_already_active(self):
-        self.test_register_201()
-        token = Token.objects.last()
-        token.user.is_active = True
-        token.user.save()
-        url = reverse('radius:phone_token_create', args=[self.default_org.slug])
-        r = self.client.post(url, HTTP_AUTHORIZATION=f'Bearer {token.key}')
-        self.assertEqual(r.status_code, 201)
-
-    @freeze_time(_TEST_DATE)
-    @capture_any_output()
-    def test_create_phone_token_400_limit_reached(self):
-        self.test_register_201()
-        token = Token.objects.last()
-        token_header = f'Bearer {token.key}'
-        max_value = app_settings.SMS_TOKEN_MAX_USER_DAILY
-        url = reverse('radius:phone_token_create', args=[self.default_org.slug])
-        for n in range(1, max_value + 2):
-            r = self.client.post(url, HTTP_AUTHORIZATION=token_header)
-            if n > max_value:
-                self.assertEqual(r.status_code, 400)
-                self.assertIn('non_field_errors', r.data)
-                self.assertIn('Maximum', str(r.data['non_field_errors']))
-            else:
-                self.assertEqual(r.status_code, 201)
-
-    @capture_any_output()
-    def test_validate_phone_token_200(self):
-        self.test_create_phone_token_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token = PhoneToken.objects.filter(user=user).last()
-        # generate entropy to ensure correct token is used
-        PhoneToken.objects.create(
-            user=user, ip=phone_token.ip, phone_number=phone_token.phone_number,
-        )
-        phone_token = PhoneToken.objects.create(
-            user=user, ip=phone_token.ip, phone_number=phone_token.phone_number
-        )
-        self.assertEqual(phone_token.attempts, 0)
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(
-            url,
-            json.dumps({'code': phone_token.token}),
-            content_type='application/json',
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 200)
-        phone_token.refresh_from_db()
-        self.assertEqual(phone_token.attempts, 1)
-        user.refresh_from_db()
-        self.assertTrue(user.is_active)
-
-    @capture_any_output()
-    def test_validate_phone_token_400_not_member(self):
-        self.test_create_phone_token_201()
-        OrganizationUser.objects.all().delete()
-        token = Token.objects.last()
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(url, HTTP_AUTHORIZATION=f'Bearer {token.key}')
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('non_field_errors', r.data)
-        self.assertIn('is not member', str(r.data['non_field_errors']))
-
-    def test_validate_phone_token_400_invalid(self):
-        self.test_create_phone_token_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token = PhoneToken.objects.filter(user=user).last()
-        self.assertEqual(phone_token.attempts, 0)
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(
-            url, {'code': '123456'}, HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 400)
-        user.refresh_from_db()
-        self.assertFalse(user.is_active)
-        phone_token.refresh_from_db()
-        self.assertEqual(phone_token.attempts, 1)
-        self.assertIn('non_field_errors', r.data)
-        self.assertIn('Invalid code', str(r.data['non_field_errors']))
-
-    @capture_any_output()
-    def test_validate_phone_token_400_expired(self):
-        self.test_create_phone_token_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token = PhoneToken.objects.filter(user=user).last()
-        phone_token.valid_until -= timedelta(days=3)
-        phone_token.save()
-        self.assertEqual(phone_token.attempts, 0)
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(
-            url,
-            {'code': phone_token.token},
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 400)
-        user.refresh_from_db()
-        self.assertFalse(user.is_active)
-        phone_token.refresh_from_db()
-        self.assertEqual(phone_token.attempts, 1)
-        self.assertIn('non_field_errors', r.data)
-        self.assertIn(
-            'This verification code has expired', str(r.data['non_field_errors'])
-        )
-
-    @capture_any_output()
-    def test_validate_phone_token_400_max_attempts(self):
-        self.test_create_phone_token_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token = PhoneToken.objects.filter(user=user).last()
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        max_value = app_settings.SMS_TOKEN_MAX_ATTEMPTS + 1
-        for n in range(1, max_value + 2):
-            r = self.client.post(
-                url, {'code': '123456'}, HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-            )
-            self.assertEqual(r.status_code, 400)
-            phone_token.refresh_from_db()
-            self.assertEqual(phone_token.attempts, n)
-            if n > max_value:
-                self.assertIn('non_field_errors', r.data)
-                self.assertIn(
-                    'Maximum number of allowed attempts',
-                    str(r.data['non_field_errors']),
-                )
-        user.refresh_from_db()
-        self.assertFalse(user.is_active)
-
-    def test_validate_phone_token_401(self):
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(url)
-        self.assertEqual(r.status_code, 401)
-
-    @capture_any_output()
-    def test_validate_phone_token_400_user_already_active(self):
-        self.test_create_phone_token_201()
-        user = User.objects.get(email=self._test_email)
-        user.is_active = True
-        user.save()
-        user_token = Token.objects.filter(user=user).last()
-        phone_token = PhoneToken.objects.filter(user=user).last()
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(
-            url,
-            {'code': phone_token.token},
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('non_field_errors', r.data)
-        self.assertIn('already active', str(r.data['non_field_errors']))
-
-    def test_validate_phone_token_400_no_token(self):
-        self.test_register_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        self.assertEqual(PhoneToken.objects.count(), 0)
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(
-            url, {'code': '123456'}, HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('non_field_errors', r.data)
-        self.assertIn(
-            'No verification code found in the system', str(r.data['non_field_errors'])
-        )
-
-    def test_validate_phone_token_400_code_blank(self):
-        self.test_register_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(
-            url, {'code': ''}, HTTP_AUTHORIZATION=f'Bearer {user_token.key}'
-        )
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('code', r.data)
-
-    def test_validate_phone_token_400_missing_code(self):
-        self.test_register_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(url, HTTP_AUTHORIZATION=f'Bearer {user_token.key}')
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('code', r.data)
-
-    @capture_any_output()
-    def test_change_phone_number_200(self):
-        self.test_create_phone_token_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token_qs = PhoneToken.objects.filter(user=user)
-        self.assertEqual(phone_token_qs.count(), 1)
-        url = reverse('radius:phone_number_change', args=[self.default_org.slug])
-        new_phone_number = '+595972157444'
-        r = self.client.post(
-            url,
-            json.dumps({'phone_number': new_phone_number}),
-            content_type='application/json',
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 200)
-        self.assertEqual(phone_token_qs.count(), 2)
-        user.refresh_from_db()
-        self.assertFalse(user.is_active)
-        self.assertEqual(user.phone_number, user.phone_number)
-
-        code = phone_token_qs.first().token
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        r = self.client.post(
-            url,
-            json.dumps({'code': code}),
-            content_type='application/json',
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 200)
-        self.assertEqual(phone_token_qs.count(), 2)
-        user.refresh_from_db()
-        self.assertTrue(user.is_active)
-        self.assertEqual(user.phone_number, new_phone_number)
-
-    def test_change_phone_number_400_same_number(self):
-        self.test_create_phone_token_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token_qs = PhoneToken.objects.filter(user=user)
-        self.assertEqual(phone_token_qs.count(), 1)
-        url = reverse('radius:phone_number_change', args=[self.default_org.slug])
-        new_phone_number = '+393664255801'
-        r = self.client.post(
-            url,
-            json.dumps({'phone_number': new_phone_number}),
-            content_type='application/json',
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 400)
-        self.assertEqual(phone_token_qs.count(), 1)
-        self.assertIn('phone_number', r.data)
-        self.assertIn('must be different', str(r.data['phone_number']))
-
-    def test_change_phone_number_400_not_blank(self):
-        self.test_create_phone_token_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token_qs = PhoneToken.objects.filter(user=user)
-        self.assertEqual(phone_token_qs.count(), 1)
-        url = reverse('radius:phone_number_change', args=[self.default_org.slug])
-        r = self.client.post(
-            url,
-            json.dumps({'phone_number': ''}),
-            content_type='application/json',
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 400)
-        self.assertEqual(phone_token_qs.count(), 1)
-        self.assertIn('phone_number', r.data)
-        self.assertIn('not be blank', str(r.data['phone_number']))
-
-    def test_change_phone_number_400_required(self):
-        self.test_create_phone_token_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token_qs = PhoneToken.objects.filter(user=user)
-        self.assertEqual(phone_token_qs.count(), 1)
-        url = reverse('radius:phone_number_change', args=[self.default_org.slug])
-        r = self.client.post(
-            url,
-            content_type='application/json',
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 400)
-        self.assertEqual(phone_token_qs.count(), 1)
-        self.assertIn('phone_number', r.data)
-        self.assertIn('required', str(r.data['phone_number']))
-
-    def test_phone_number_restriction_in_signup(self):
-        app_settings.ALLOWED_MOBILE_PREFIXES = ['+33']
-        self.assertEqual(User.objects.all().count(), 0)
-        user_param = {
-            'username': 'test',
-            'email': 'test@email.com',
-            'password1': 'password',
-            'password2': 'password',
-        }
-        url = reverse('radius:rest_register', args=[self.default_org.slug])
-        msg = 'This international mobile prefix is not allowed.'
-
-        with self.subTest('test create user with number allowed at global level'):
-            user_param.update({'phone_number': '+33675579231'})
-            r = self.client.post(url, user_param)
-            self.assertEqual(r.status_code, 201)
-            self.assertEqual(User.objects.all().count(), 1)
-            self.assertIn('key', r.data)
-
-        with self.subTest('test create user with number not allowed at org level'):
-            user_param.update(
-                {
-                    'email': 'test1@email.com',
-                    'username': 'test1',
-                    'phone_number': '+44 7795 106991',
-                }
-            )
-            r = self.client.post(url, user_param)
-            self.assertEqual(r.status_code, 400)
-            self.assertIn(msg, str(r.data['phone_number']))
-            self.assertEqual(User.objects.all().count(), 1)
-
-        with self.subTest('test create user with number allowed at org level'):
-            radius_settings = self.default_org.radius_settings
-            radius_settings.allowed_mobile_prefixes = '+1,+44'
-            radius_settings.full_clean()
-            radius_settings.save()
-            user_param.update(
-                {
-                    'email': 'test1@email.com',
-                    'username': 'test1',
-                    'phone_number': '+44 7795 106991',
-                }
-            )
-            r = self.client.post(url, user_param)
-            self.assertEqual(r.status_code, 201)
-            self.assertEqual(User.objects.all().count(), 2)
-            self.assertIn('key', r.data)
-
-        with self.subTest('test create user with number not allowed'):
-            user_param.update(
-                {
-                    'email': 'test2@email.com',
-                    'username': 'test2',
-                    'phone_number': '+237675752913',
-                }
-            )
-            r = self.client.post(url, user_param)
-            self.assertEqual(r.status_code, 400)
-            self.assertIn(msg, str(r.data['phone_number']))
-            self.assertEqual(User.objects.all().count(), 2)
-
-        app_settings.ALLOWED_MOBILE_PREFIXES = []
-
-    @capture_any_output()
-    def test_change_phone_number_restriction(self):
-        self.test_register_201()
-        app_settings.ALLOWED_MOBILE_PREFIXES = ['+33']
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token_qs = PhoneToken.objects.filter(user=user)
-        url = reverse('radius:phone_number_change', args=[self.default_org.slug])
-        msg = 'This international mobile prefix is not allowed.'
-
-        self.assertEqual(phone_token_qs.count(), 0)
-
-        with self.subTest('test change number not allowed at global level'):
-            r = self.client.post(
-                url,
-                json.dumps({'phone_number': '+237675579231'}),
-                content_type='application/json',
-                HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-            )
-            self.assertEqual(r.status_code, 400)
-            self.assertIn(msg, str(r.data['phone_number']))
-            self.assertEqual(phone_token_qs.count(), 0)
-
-        with self.subTest('test change number allowed at global level'):
-            phone_number = '+33675579245'
-            r = self.client.post(
-                url,
-                json.dumps({'phone_number': phone_number}),
-                content_type='application/json',
-                HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-            )
-            self.assertEqual(r.status_code, 200)
-            self.assertEqual(phone_token_qs.count(), 1)
-            code = phone_token_qs.get(user=user, phone_number=phone_number).token
-            r = self.client.post(
-                reverse('radius:phone_token_validate', args=[self.default_org.slug]),
-                json.dumps({'code': code}),
-                content_type='application/json',
-                HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-            )
-            self.assertEqual(r.status_code, 200)
-            user.refresh_from_db()
-            self.assertEqual(user.phone_number, phone_number)
-
-        with self.subTest('test change number not allowed at org level'):
-            r = self.client.post(
-                url,
-                json.dumps({'phone_number': '+44 7795 106991'}),
-                content_type='application/json',
-                HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-            )
-            self.assertEqual(r.status_code, 400)
-            self.assertIn(msg, str(r.data['phone_number']))
-            self.assertEqual(phone_token_qs.count(), 1)
-
-        with self.subTest('test change number allowed at org level'):
-            radius_settings = self.default_org.radius_settings
-            radius_settings.allowed_mobile_prefixes = '+1,+44'
-            radius_settings.full_clean()
-            radius_settings.save()
-            phone_number = '+44 7795 106991'
-            r = self.client.post(
-                url,
-                json.dumps({'phone_number': phone_number}),
-                content_type='application/json',
-                HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-            )
-            self.assertEqual(r.status_code, 200)
-            self.assertEqual(phone_token_qs.count(), 2)
-            code = phone_token_qs.get(user=user, phone_number=phone_number).token
-            r = self.client.post(
-                reverse('radius:phone_token_validate', args=[self.default_org.slug]),
-                json.dumps({'code': code}),
-                content_type='application/json',
-                HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-            )
-            self.assertEqual(r.status_code, 200)
-            user.refresh_from_db()
-            self.assertEqual(user.phone_number, phone_number)
-
-        app_settings.ALLOWED_MOBILE_PREFIXES = []
-
-    def _test_change_phone_number_sms_on_helper(self, is_active):
-        self.test_register_201()
-        user = User.objects.get(email=self._test_email)
-        user_token = Token.objects.filter(user=user).last()
-        phone_token_qs = PhoneToken.objects.filter(user=user)
-        url = reverse('radius:phone_number_change', args=[self.default_org.slug])
-        old_phone_number = '+237675582041'
-        user.phone_number = old_phone_number
-        user.is_active = is_active
-        user.save()
-        new_phone_number = '+237675582042'
-
-        self.assertEqual(phone_token_qs.count(), 0)
-
-        r = self.client.post(
-            url,
-            json.dumps({'phone_number': new_phone_number}),
-            content_type='application/json',
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 200)
-        self.assertEqual(phone_token_qs.count(), 1)
-        self.assertEqual(phone_token_qs.first().phone_number, new_phone_number)
-        user.refresh_from_db()
-        self.assertEqual(user.phone_number, old_phone_number)
-        self.assertFalse(user.is_active)
-
-    @capture_any_output()
-    def test_active_user_change_phone_number_sms_on(self):
-        self._test_change_phone_number_sms_on_helper(True)
-
-    @capture_any_output()
-    def test_inactive_user_change_phone_number_sms_on(self):
-        self._test_change_phone_number_sms_on_helper(False)
-
-    def _create_user_helper(self, options):
-        self._superuser_login()
-        url = reverse('radius:rest_register', args=[self.default_org.slug])
-        r = self.client.post(url, options)
-        self.assertEqual(r.status_code, 201)
-        url = reverse('radius:phone_token_create', args=[self.default_org.slug])
-        r1 = self.client.post(
-            url,
-            content_type='application/json',
-            HTTP_AUTHORIZATION=f'Bearer {r.data["key"]}',
-        )
-        self.assertEqual(r1.status_code, 201)
-
-    def _test_phone_number_unique_helper(self, phone_number):
-        user = User.objects.get(email='user2@gmail.com')
-        user_token = Token.objects.filter(user=user).last()
-        phone_token_qs = PhoneToken.objects.filter(user=user)
-        url = reverse('radius:phone_number_change', args=[self.default_org.slug])
-        self.assertEqual(phone_token_qs.count(), 1)
-
-        r = self.client.post(
-            url,
-            json.dumps({'phone_number': phone_number}),
-            content_type='application/json',
-            HTTP_AUTHORIZATION=f'Bearer {user_token.key}',
-        )
-        self.assertEqual(r.status_code, 400)
-        self.assertIn('phone_number', r.data)
-        self.assertEqual(phone_token_qs.count(), 1)
-
-    @capture_any_output()
-    def test_phone_token_phone_number_unique(self):
-        self._create_user_helper(
-            {
-                'username': 'user1',
-                'email': 'user1@gmail.com',
-                'password1': 'password',
-                'password2': 'password',
-                'phone_number': '+237678879231',
-            }
-        )
-        self._create_user_helper(
-            {
-                'username': 'user2',
-                'email': 'user2@gmail.com',
-                'password1': 'password',
-                'password2': 'password',
-                'phone_number': '+237674479231',
-            }
-        )
-        self._test_phone_number_unique_helper('+237678879231')
-
-    @capture_any_output()
-    def test_user_phone_number_unique(self):
-        User.objects.create_user(
-            username='user1',
-            email='email1@gmail.com',
-            password='pass',
-            phone_number='+23767779235',
-        )
-        self._create_user_helper(
-            {
-                'username': 'user2',
-                'email': 'user2@gmail.com',
-                'password1': 'password',
-                'password2': 'password',
-                'phone_number': '+237674479231',
-            }
-        )
-        user = User.objects.get(email='user2@gmail.com')
-        user.is_active = True
-        user.save()
-        user.refresh_from_db()
-        # Testing for Phone token validation error due to same phone number ,
-        self._test_phone_number_unique_helper('+23767779235')
-        # is_active state of user should not change because an error
-        # occurred during phone token creation.
-        self.assertTrue(user.is_active)
-
-
-class TestIsSmsVerificationEnabled(ApiTokenMixin, BaseTestCase):
-
-    _test_email = 'test@openwisp.org'
-
-    def setUp(self):
-        super().setUp()
-        radius_settings = self.default_org.radius_settings
-        radius_settings.sms_verification = False
-        radius_settings.full_clean()
-        radius_settings.save()
-
-    def test_register_201_phone_number_empty(self):
-        self._superuser_login()
-        self.assertEqual(User.objects.count(), 1)
-        url = reverse('radius:rest_register', args=[self.default_org.slug])
-        phone_number = '+393664255801'
-        r = self.client.post(
-            url,
-            {
-                'username': self._test_email,
-                'email': self._test_email,
-                'password1': 'password',
-                'password2': 'password',
-                'phone_number': phone_number,
-            },
-        )
-        self.assertEqual(r.status_code, 201)
-        self.assertIn('key', r.data)
-        self.assertEqual(User.objects.count(), 2)
-        user = User.objects.get(email=self._test_email)
-        self.assertTrue(user.is_member(self.default_org))
-        self.assertEqual(user.phone_number, None)
-        self.assertTrue(user.is_active)
-
-    @capture_stderr()
-    def test_create_phone_token_403(self):
-        url = reverse('radius:phone_token_create', args=[self.default_org.slug])
-        with self.assertNumQueries(1):
-            r = self.client.post(url)
-        self.assertEqual(r.status_code, 403)
-        self.assertIn('SMS verification is not enabled', str(r.data))
-
-    @capture_stderr()
-    def test_validate_phone_token_403(self):
-        url = reverse('radius:phone_token_validate', args=[self.default_org.slug])
-        with self.assertNumQueries(1):
-            r = self.client.post(url)
-        self.assertEqual(r.status_code, 403)
-        self.assertIn('SMS verification is not enabled', str(r.data))
-
-    @capture_stderr()
-    def test_change_phone_number_403(self):
-        url = reverse('radius:phone_number_change', args=[self.default_org.slug])
-        with self.assertNumQueries(1):
-            r = self.client.post(url)
-        self.assertEqual(r.status_code, 403)
-        self.assertIn('SMS verification is not enabled', str(r.data))
-
-    @capture_stderr()
-    def test_missing_radius_settings(self):
-        self.default_org.radius_settings.delete()
-        url = reverse('radius:phone_token_create', args=[self.default_org.slug])
-        with self.assertNumQueries(1):
-            r = self.client.post(url)
-        self.assertEqual(r.status_code, 500)
-        self.assertIn('Could not complete operation', str(r.data))
