+++ tests/openwisp2/sample_radius/api/views.py
@@ -1,5 +1,6 @@
+from openwisp_radius.api.freeradius_views import AuthorizeView as BaseAuthorizeView
+from openwisp_radius.api.freeradius_views import PostAuthView as BasePostAuthView
 from openwisp_radius.api.views import AccountingView as BaseAccountingView
-from openwisp_radius.api.views import AuthorizeView as BaseAuthorizeView
 from openwisp_radius.api.views import BatchView as BaseBatchView
 from openwisp_radius.api.views import ChangePhoneNumberView as BaseChangePhoneNumberView
 from openwisp_radius.api.views import CreatePhoneTokenView as BaseCreatePhoneTokenView
@@ -12,7 +13,6 @@ from openwisp_radius.api.views import (
     PasswordResetConfirmView as BasePasswordResetConfirmView,
 )
 from openwisp_radius.api.views import PasswordResetView as BasePasswordResetView
-from openwisp_radius.api.views import PostAuthView as BasePostAuthView
 from openwisp_radius.api.views import RegisterView as BaseRegisterView
 from openwisp_radius.api.views import UserAccountingView as BaseUserAccountingView
 from openwisp_radius.api.views import ValidateAuthTokenView as BaseValidateAuthTokenView
